Add files via upload
This is the code i attempted to write that connects the user registration to our database.  the "register.php" file is already pointed (action) towards the process_registration.php.  My code must be a little off though because registering info still doesnt transfer to the database.  Feel free to try figuring out what is wrong.